Port desktop app on all platforms to Qt 6.2.x

See merge request ws/client/desktop/client-desktop-public!52

+++ client/gui/backend/backend.h
@@ -105,7 +105,7 @@ public:
     void abortInitialization();
 
     void sendUpdateWindowInfo(qint32 mainWindowCenterX, qint32 mainWindowCenterY);
-    void sendUpdateVersion(qint32 mainWindowHandle);
+    void sendUpdateVersion(qint64 mainWindowHandle);
     void cancelUpdateVersion();
 
     void sendMakeHostsFilesWritableWin();
